data-pipeline/ingestion/locations-and-songs/requirements.txt
google-cloud-storage
pyarrow
requests
